backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
tensorflow
Pillow
numpy
